Add files via upload
The main idea of this project entitled Human-Machine Activity Monitoring using Computer Vision is to design a system using convolutional neural networks for facilitating human supervision in monitors the activity of humans and machines in a warehouse. This will in return improve the resource utilization/management as well. Vision-based Human-Machine activity monitoring is important as it takes care of one of the major problems faced by industries. In every organization, there is a surveillance room where humans supervise the activities inside the warehouse. It is scientifically proven that the attention span of a human is limited and is greatly affected by environmental factors and repetitive tasks. A viable solution would be to design an intelligent system that will help make the surveillance task easier. Our prime focus is to facilitate the existing industrial environment rather than incorporating something that needs large-scale changes because such changes are not usually accepted at once in our local industry. Organizations are becoming data-driven. We would use the already present surveillance cameras present in warehouses and develop a program that will detect the efficiency of the workers throughout the workday. The system is not in any way going to replace the human supervisor but will facilitate him with an effortlessly generated report and relatively better observation. The industry will have continuously updated data records what will lead to better resource management.